+++ .env
@@ -1,22 +1,37 @@
-# DATA_SIZE=1k
-# CHUNK_SIZE=200
+#======= General settings =======#
+PROJECT_ROOT=${HOME}/code/Wikar235/disease-recognition
+DATASET=lokisilvres/dental-disease-panoramic-detection-dataset
+DATA_FILE=${HOME}/.cache/kagglehub/datasets/lokisilvres/dental-disease-panoramic-detection-dataset/versions/6/YOLO/YOLO/data.yaml
+MODEL_STORAGE="local" # "local", "gcs", or "mlflow"
+LOCAL_REGISTRY_PATH=${PROJECT_ROOT}/models
+SCRIPT_DIR=${PROJECT_ROOT}/scripts
+WORK_DIR=${PROJECT_ROOT}/.tmp
 
-# MODEL
-# MODEL_PATH=
+#======= Variables for training =======#
+WEIGHTS_TRAIN=yolov8n.pt
+WEIGHTS_PREDICT=trained_model_100epoch.pt
+EPOCHS=100
+BATCH_SIZE=16
+IMG_SIZE=640
+DEVICE=cpu
+LEARNING_RATE=0.01
+MOMENTUM=0.937
+WEIGHT_DECAY=0.0005
+AUGMENTATION=True
+#VALIDATION_SPLIT=0.2
 
-
-#TEST
-SAMPLE_IMAGE=test/0a51382f-ZAREII_HAMDOLAH_2020-07-19122324_jpg.rf.2990610a8272e47803baaac304786eff.jpg
-
-
-# GCP PROJECT & REGION
+#======= Variables for GCP =======#
+BUCKET_NAME=disease-recognition
 GCP_PROJECT=disease-recognition-470203
 GCP_REGION=us-west1
-
-
-#Docker
+BUCKET_NAME=disease-recognition
 GAR_IMAGE=dental-recognition-api
 GAR_MEMORY=2Gi
 GAR_PATH=${GCP_REGION}-docker.pkg.dev/${GCP_PROJECT}/dental-recognition/${GAR_IMAGE}:prod
 
-#SERVICE_URL=
+#======= Variables for MLflow =======#
+MLFLOW_TRACKING_URI=https://mlflow.nasebanal.com
+# MLFLOW_TRACKING_URI=https://mlflow.lewagon.ai
+MLFLOW_EXPERIMENT=disease-recognition
+MLFLOW_MODEL_NAME=disease-recognition-model
+MLFLOW_RUN_NAME=disease-recognition-run
