--- .env.sample
@@ -1,36 +1,0 @@
-#======= General settings =======#
-PROJECT_ROOT=${HOME}/code/Wikar235/disease-recognition
-DATASET=lokisilvres/dental-disease-panoramic-detection-dataset
-DATA_FILE=${HOME}/.cache/kagglehub/datasets/lokisilvres/dental-disease-panoramic-detection-dataset/versions/6/YOLO/YOLO/data.yaml
-MODEL_STORAGE="local" # "local", "gcs", or "mlflow"
-LOCAL_REGISTRY_PATH=${PROJECT_ROOT}/models
-SCRIPT_DIR=${PROJECT_ROOT}/scripts
-WORK_DIR=${PROJECT_ROOT}/.tmp
-
-#======= Variables for training =======#
-WEIGHTS_TRAIN=yolov8n.pt
-WEIGHTS_PREDICT=trained_model_100epoch.pt
-EPOCHS=100
-BATCH_SIZE=16
-IMG_SIZE=640
-DEVICE=cpu
-LEARNING_RATE=0.01
-MOMENTUM=0.937
-WEIGHT_DECAY=0.0005
-AUGMENTATION=True
-#VALIDATION_SPLIT=0.2
-
-#======= Variables for GCP =======#
-BUCKET_NAME=disease-recognition
-GCP_PROJECT=<PROJECT_ID>
-GCP_REGION=<REGION>
-BUCKET_NAME=disease-recognition
-# GAR_IMAGE=gcr.io/<PROJECT_ID>/disease-recognition:latest
-# GAR_MEMORY=32Gi
-
-#======= Variables for MLflow =======#
-MLFLOW_TRACKING_URI=https://mlflow.nasebanal.com
-# MLFLOW_TRACKING_URI=https://mlflow.lewagon.ai
-MLFLOW_EXPERIMENT=disease-recognition
-MLFLOW_MODEL_NAME=disease-recognition-model
-MLFLOW_RUN_NAME=disease-recognition-run
