+++ .env.yaml
@@ -3,6 +3,3 @@ GCP_REGION: "us-west1"
 
 GAR_IMAGE: "dental-recognition-api"
 GAR_MEMORY: "2Gi"
-
-SAMPLE_IMAGE: "test/0a51382f-ZAREII_HAMDOLAH_2020-07-19122324_jpg.rf.2990610a8272e47803baaac304786eff.jpg"
-#SERVICE_URL:
