--- src/models/params.py
@@ -1,40 +1,0 @@
-import os
-import numpy as np
-
-import dotenv
-dotenv.load_dotenv()
-
-##################  VARIABLES  ##################
-SAMPLE_IMAGE = os.environ.get("SAMPLE_IMAGE", "test/0a51382f-ZAREII_HAMDOLAH_2020-07-19122324_jpg.rf.2990610a8272e47803baaac304786eff.jpg")
-
-GCP_PROJECT = os.environ.get("GCP_PROJECT")
-GCP_REGION = os.environ.get("GCP_REGION")
-GAR_IMAGE = os.environ.get("GAR_IMAGE")
-GAR_MEMORY = os.environ.get("GAR_MEMORY")
-
-DTYPES_PROCESSED = np.float32
-
-
-# DATA_SIZE = os.environ.get("DATA_SIZE")
-# CHUNK_SIZE = int(os.environ.get("CHUNK_SIZE"))
-# MODEL_TARGET = os.environ.get("MODEL_TARGET")
-# BQ_DATASET = os.environ.get("BQ_DATASET")
-# BQ_REGION = os.environ.get("BQ_REGION")
-# BUCKET_NAME = os.environ.get("BUCKET_NAME")
-# INSTANCE = os.environ.get("INSTANCE")
-# MLFLOW_TRACKING_URI = os.environ.get("MLFLOW_TRACKING_URI")
-# MLFLOW_EXPERIMENT = os.environ.get("MLFLOW_EXPERIMENT")
-# MLFLOW_MODEL_NAME = os.environ.get("MLFLOW_MODEL_NAME")
-# PREFECT_FLOW_NAME = os.environ.get("PREFECT_FLOW_NAME")
-# PREFECT_LOG_LEVEL = os.environ.get("PREFECT_LOG_LEVEL")
-# EVALUATION_START_DATE = os.environ.get("EVALUATION_START_DATE")
-################## VALIDATIONS #################
-
-# def validate_env_value(env, valid_options):
-#     env_value = os.environ[env]
-#     if env_value not in valid_options:
-#         raise NameError(f"Invalid value for {env} in `.env` file: {env_value} must be in {valid_options}")
-
-
-# for env, valid_options in env_valid_options.items():
-#     validate_env_value(env, valid_options)
